requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
requests
google-auth>=2.23.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
typing-extensions>=4.8.0
tenacity
aiohttp
google-api-core
google-cloud-core
protobuf
grpcio
googleapis-common-protos
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
typing-extensions>=4.8.0
requests>=2.31.0
six>=1.16.0
cachetools>=5.0.0
rsa>=4.7
urllib3
certifi
charset-normalizer